transferred office hours functions to the bio sci scraper
